# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Jobbly API
  description: >
    Backend de reclutamiento Jobbly (Go + PostgreSQL). Contrato consumido por
    el frontend RECLUTAMIENTO-AI. La sesión usa access token JWT (Bearer, 15
    min) + refresh token rotativo en cookie httpOnly `jobbly_refresh`
    (Path=/api/v1/auth).
  version: "0.1.0"
servers:
  - url: http://localhost:4000
    description: Desarrollo local
  - url: https://api.jobbly.mx
    description: Producción (VPS)
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Iniciar sesión
      description: >
        Rate limit estricto por IP (10/min). Si `recruiter.passwordChanged`
        es false, el cliente debe forzar el cambio de contraseña antes de
        usar el resto del API (los endpoints protegidos responden 403
        `password_change_required`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Sesión creada. Set-Cookie `jobbly_refresh` (httpOnly).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Renovar sesión (rotación de refresh token)
      description: >
        Usa la cookie `jobbly_refresh`. El token presentado se revoca y se
        emite uno nuevo (rotación). Presentar un token ya revocado se trata
        como robo y revoca todas las sesiones del usuario.
      responses:
        "200":
          description: Nueva sesión. Set-Cookie con el refresh rotado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Perfil del reclutador autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  recruiter: { $ref: "#/components/schemas/Recruiter" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/auth/forgot-password:
    post:
      tags: [auth]
      summary: Solicitar restablecimiento de contraseña
      description: >
        Rate limit estricto (10/5s por IP). Siempre responde 200 para no
        revelar si el email existe (anti-enumeración). Si el reclutador
        existe, se envía un email con un enlace de un solo uso válido 30 min.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Solicitud procesada (independientemente de si el email existe).
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/reset-password:
    post:
      tags: [auth]
      summary: Restablecer contraseña con token de un solo uso
      description: >
        Rate limit estricto (10/5s por IP). El token viene del enlace del
        email (query param `token`). Solo válido 30 min y un único uso.
        Al completar, todas las sesiones previas quedan revocadas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, description: "Token opaco recibido por email" }
                newPassword: { type: string, minLength: 10 }
      responses:
        "200":
          description: Contraseña restablecida. El usuario debe iniciar sesión.
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: Cambiar contraseña (obligatorio si es temporal)
      description: >
        Valida la contraseña actual, aplica la política (≥10 caracteres,
        letras y números), marca `passwordChanged=true` y revoca todas las
        sesiones anteriores emitiendo una nueva.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 10 }
      responses:
        "200":
          description: Contraseña cambiada; nueva sesión emitida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/recruiters:
    get:
      tags: [recruiters]
      summary: Listar reclutadores de la organización
      description: Requiere rol Administrador y contraseña definitiva.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  recruiters:
                    type: array
                    items: { $ref: "#/components/schemas/Recruiter" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [recruiters]
      summary: Dar de alta un reclutador
      description: >
        Solo Administradores. El nuevo reclutador nace con
        `passwordChanged=false` y debe cambiar la contraseña temporal en su
        primer login.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, temporaryPassword, permission]
              properties:
                name: { type: string, description: "Solo letras y espacios" }
                email: { type: string, format: email }
                phone: { type: string, description: "Dígitos, + y espacios" }
                temporaryPassword: { type: string, minLength: 8 }
                permission:
                  type: string
                  enum: [Administrador, Ejecutivo]
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  recruiter: { $ref: "#/components/schemas/Recruiter" }
        "400": { $ref: "#/components/responses/Error" }
        "409":
          description: Email ya registrado (`email_taken`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/v1/recruiters/{id}:
    patch:
      tags: [recruiters]
      summary: Activar / desactivar un reclutador
      description: >
        Solo Administradores. Desactivar revoca las sesiones abiertas del
        usuario. No es posible desactivar la propia cuenta
        (`self_deactivation`).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive: { type: boolean }
      responses:
        "200":
          description: ok
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/companies:
    get:
      tags: [companies]
      summary: Listar compañías con conteo de vacantes (solo Administrador)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: ok }
    post:
      tags: [companies]
      summary: Crear compañía (solo Administrador)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 2, maxLength: 120 }
      responses:
        "201": { description: creada }
        "409": { description: nombre duplicado (`name_taken`) }
  /api/v1/companies/{id}:
    patch:
      tags: [companies]
      summary: Renombrar compañía (solo Administrador)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: ok }
  /api/v1/vacancies:
    get:
      tags: [vacancies]
      summary: Listar vacantes de la organización (filtros y paginación server-side)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [draft, published, closed] } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: "{ vacancies: Vacancy[], total: int }"
    post:
      tags: [vacancies]
      summary: Crear vacante (borrador o publicada)
      description: >
        `company` por nombre hace find-or-create en la organización. Con
        `publish: true` nace en estado `published` con `publishedAt`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVacancyInput" }
      responses:
        "201":
          description: Creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  vacancy: { $ref: "#/components/schemas/Vacancy" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/vacancies/{id}:
    get:
      tags: [vacancies]
      summary: Obtener vacante de la organización
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: ok }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [vacancies]
      summary: Cambiar estado (ciclo draft→published→closed; closed→published reabre)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [draft, published, closed] }
      responses:
        "200": { description: ok }
        "400": { description: transición inválida (`bad_status`) }
...
  /api/v1/public/vacancies/{id}:
    get:
      tags: [public]
      summary: Vacante publicada por id (sin auth)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: ok }
        "404": { description: no existe o no está publicada }
  /api/v1/public/applications:
    post:
      tags: [public]
      summary: Postulación pública a una vacante (sin auth, rate limit 10/min por IP)
      description: >
        Upsert del candidato por (organización, teléfono): la postulación más
        reciente actualiza el perfil. Una postulación por candidato por vacante.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyInput" }
      responses:
        "201": { description: "{ applicationId: uuid }" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { description: vacante no disponible (`vacancy_not_open`) }
        "409": { description: ya postulado (`already_applied`) }
  /api/v1/candidates:
    get:
      tags: [candidates]
      summary: Listar candidatos de la organización (búsqueda y paginación)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
      responses:
        "200": { description: "{ candidates: Candidate[], total: int }" }
  /api/v1/applications:
    get:
      tags: [candidates]
      summary: Listar postulaciones (opcionalmente por vacante)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: vacancyId, in: query, schema: { type: string, format: uuid } }
      responses:
        "200": { description: "{ applications: Application[] }" }
  /api/v1/applications/{id}:
    patch:
      tags: [candidates]
      summary: Cambiar estado de la postulación (pipeline)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [nueva, en_revision, entrevista, rechazada, contratada]
      responses:
        "200": { description: ok }
  /api/v1/rutas:
    get:
      tags: [rutas]
      summary: Listar rutas de transporte con sus vacantes vinculadas
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: "{ rutas: Ruta[] }" }
    post:
      tags: [rutas]
      summary: Crear ruta (vacancyIds valida pertenencia a la organización)
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: creada }
  /api/v1/rutas/{id}:
    patch:
      tags: [rutas]
      summary: Actualizar ruta y sus vínculos
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: ok }
    delete:
      tags: [rutas]
      summary: Eliminar ruta
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: ok }
  /api/v1/summary:
    get:
      tags: [ops]
      summary: Métricas de negocio de la organización autenticada
      description: >
        Devuelve conteos reales de candidatos, vacantes (por estado), compañías,
        reclutadores activos y pipeline de postulaciones. Requiere sesión activa.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary: { $ref: "#/components/schemas/OrgSummary" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/facebookads/config:
    get:
      tags: [facebookads]
      summary: Obtener configuración de Meta Ads (sin exponer secretos)
      description: >
        Devuelve la configuración de la organización. Los campos `appSecretSet`
        y `accessTokenSet` indican si hay un valor cifrado; los textos planos
        nunca se exponen. Si no existe configuración, devuelve valores por defecto.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  config: { $ref: "#/components/schemas/FbConfig" }
        "401": { $ref: "#/components/responses/Error" }
    put:
      tags: [facebookads]
      summary: Guardar configuración de Meta Ads (upsert)
      description: >
        Upsert de la configuración. Campos `appSecret` y `accessToken` vacíos
        conservan el valor cifrado anterior. Guardar una nueva config resetea
        `isConnected=false`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FbSaveConfigInput" }
      responses:
        "200":
          description: Guardado
          content:
            application/json:
              schema:
                type: object
                properties:
                  config: { $ref: "#/components/schemas/FbConfig" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/facebookads/test:
    post:
      tags: [facebookads]
      summary: Probar conexión con Meta Graph API
      description: >
        Llama GET /act_{adAccountId} desde el servidor y actualiza `isConnected`
        según el resultado. Requiere access token y ad_account_id configurados.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Conexión exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountName: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/facebookads/oauth/url:
    get:
      tags: [facebookads]
      summary: Generar URL de autorización OAuth de Meta
      description: >
        Construye la URL del diálogo de OAuth de Facebook con CSRF state (10 min
        TTL en memoria). El frontend redirige el navegador a esta URL para
        iniciar el flujo SSO.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
        "400":
          description: App ID no configurado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/facebookads/oauth/callback:
    get:
      tags: [facebookads]
      summary: Callback OAuth de Meta (público — el browser llega aquí)
      description: >
        Meta redirige el navegador a este endpoint tras la autorización. Valida
        el CSRF state, intercambia el código por un long-lived token (60 días) y
        lo cifra con AES-GCM en DB. Redirige al frontend con `?fb=connected` o
        `?fb=error`. **No requiere JWT** (es un redirect del browser).
      parameters:
        - { name: code,  in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        "302":
          description: Redirect al frontend con ?fb=connected o ?fb=error
  /api/v1/facebookads/accounts:
    get:
      tags: [facebookads]
      summary: Listar cuentas publicitarias accesibles vía el token actual
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items: { $ref: "#/components/schemas/FbAdAccount" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/facebookads/pages:
    get:
      tags: [facebookads]
      summary: Listar páginas de Facebook administradas por el usuario del token
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: array
                    items: { $ref: "#/components/schemas/FbPage" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/facebookads/selection:
    put:
      tags: [facebookads]
      summary: Guardar cuenta publicitaria y página seleccionadas
      description: >
        Actualiza `ad_account_id` y `page_id` sin resetear `isConnected`.
        La conexión OAuth sigue válida.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adAccountId, pageId]
              properties:
                adAccountId: { type: string }
                pageId: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  config: { $ref: "#/components/schemas/FbConfig" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { description: configuración no encontrada }
  /api/v1/facebookads/disconnect:
    post:
      tags: [facebookads]
      summary: Desconectar Meta — borra el access token y resetea el estado
      description: >
        Limpia `access_token_enc`, `ad_account_id`, `page_id` y pone
        `is_connected=false`. El App ID y App Secret se conservan.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error con sobre uniforme
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    Recruiter:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        permission: { type: string, enum: [Administrador, Ejecutivo] }
        isActive: { type: boolean }
        passwordChanged: { type: boolean }
        createdBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      properties:
        accessToken: { type: string, description: "JWT Bearer, 15 minutos" }
        recruiter: { $ref: "#/components/schemas/Recruiter" }
...
    CreateVacancyInput:
      type: object
      required: [surveyName, jobTitle, jobDescription, company, workMode, requestedSex, activities]
      properties:
        surveyName: { type: string, maxLength: 200 }
        jobTitle: { type: string, maxLength: 200 }
        jobDescription: { type: string }
        company: { type: string, description: "find-or-create por nombre" }
        state: { type: string }
        municipality: { type: string }
        workMode: { type: string, enum: [Hibrida, Presencial, Remoto] }
        salaryRange: { type: string }
        schedule: { type: string }
        requestedSex: { type: string, enum: [Hombre, Mujer, Ambos] }
        educationLevel: { type: array, items: { type: string } }
        activities: { type: array, items: { type: string }, minItems: 1 }
        customBoxes: { type: array, items: { type: string } }
        publish: { type: boolean, default: false }
...
    FbAdAccount:
      type: object
      properties:
        id:     { type: string }
        name:   { type: string }
        status: { type: integer, description: "1=activa, 2=desactivada, 3=no pagada, etc." }
...